refactoring: move App component into containers
