+++ src/containers/App/App.js
@@ -4,9 +4,9 @@ require('./App.css')
 import React from 'react'
 import Navbar from './Navbar/Navbar'
 import { Switch, Route } from 'react-router-dom'
-import Shop from '../containers/Shop/Shop'
-import ShoppingCart from '../containers/ShoppingCart/ShoppingCart'
-import ProductDetails from '../containers/ProductDetails/ProductDetails'
+import Shop from '../../containers/Shop/Shop'
+import ShoppingCart from '../../containers/ShoppingCart/ShoppingCart'
+import ProductDetails from '../../containers/ProductDetails/ProductDetails'
 
 //let yeomanImage = require('../../images/yeoman.png');
